refactor: replace FailedSetupVaultKeygenStep with FlowErrorPageContent in KeygenFlow
- Removed the `FailedSetupVaultKeygenStep` component and integrated `FlowErrorPageContent` for improved error handling in the key generation flow.
- Updated the error rendering logic to utilize `extractErrorMsg` for better error message extraction and display.
- Refactored `SetupVaultEducationSlides` to use `CreateVaultDynamicEducationContent`, simplifying the education slide component structure and enhancing maintainability.
- Deleted unused `DynamicEducationalContent` component to streamline the codebase.
