+++ clients/desktop/src/vault/setup/shared/SetupVaultCreationStep/SetupVaultEducationSlides/index.tsx
@@ -1,31 +1,15 @@
 import { KeygenStep } from '@core/mpc/keygen/KeygenStep'
-import { useStepNavigation } from '@lib/ui/hooks/useStepNavigation'
+import { CreateVaultDynamicEducationContent } from '@core/ui/mpc/keygen/create/education/CreateVaultDynamicEducationalContent'
 import { VStack } from '@lib/ui/layout/Stack'
 import { PageContent } from '@lib/ui/page/PageContent'
 import { ValueProp } from '@lib/ui/props'
-import { useInterval } from 'react-use'
 import styled from 'styled-components'
 
-import { DynamicEducationContent } from './components/DynamicEducationalContent'
 import { SlidesLoader } from './components/SlidesLoader'
-
-const SLIDE_DURATION_IN_MS = 6000
-const steps = [
-  'multiFactor',
-  'selfCustodial',
-  'crossChain',
-  'availablePlatforms',
-  'seedlessWallet',
-] as const
-
-export type SetupFastVaultEducationSlidesStep = (typeof steps)[number]
 
 export const SetupVaultEducationSlides = ({
   value,
 }: ValueProp<KeygenStep | null>) => {
-  const { step, toNextStep } = useStepNavigation({ steps, circular: true })
-  useInterval(() => toNextStep(), SLIDE_DURATION_IN_MS)
-
   return (
     <PageContent>
       <Wrapper
@@ -35,7 +19,7 @@ export const SetupVaultEducationSlides = ({
         gap={48}
       >
         <VStack flexGrow justifyContent="center">
-          <DynamicEducationContent value={step} />
+          <CreateVaultDynamicEducationContent />
         </VStack>
         <SlidesLoader value={value} />
       </Wrapper>
